fix: move method to block
